Added a block iterator for createSTEMImage
This block iterator is an input iterator that iterates
over blocks one at a time from the list of files it has been
given. This allows the program to avoid loading all of the
data in at once, and instead, each block is loaded when it is
needed.

Some work probably needs to be done. I'm not convinced that
my design of the BlockIterator class is the best.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>