Allow StreamReader to read more than one file
This commit allows a single StreamReader to read multiple
files. This will be useful when we introduce an iterator for
the StreamReader.

Now, an EoFException is only thrown if we try to read a block
of data and we are at the EoF of the final file. If we have not
reached the final file, it will simply move onto the next file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>