Use a StreamReader::iterator instead of BlockIt
This adds a StreamReader::iterator class that attempts to
obey all of the rules for input iterators. This also replaces
the use of BlockIterator with the StreamReader::iterator.

It does not run in python, unfortunately. This error occurs
when we attempt to run it:

```
Traceback (most recent call last):
  File "stem_image.py", line 25, in <module>
    img = image.create_stem_image(reader, 160, 160,  40, 288);
  File "/home/patrick/virtualenvs/stempy/lib/python3.6/site-packages/stempy/image/__init__.py", line 7, in create_stem_image
    img =  _image.create_stem_image(reader.begin(), reader.end(),
TypeError: Unable to convert function return value to a Python type! The signature was
	(self: stempy._io._reader) -> stempy::StreamReader::iterator
```

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>