Fix two bugs involving StreamReader::iterator
1. The first `block` needs to be read when StreamReader
gets constructed, so that the StreamReader::iterator
`first` will have access to it via `*first` or `first->`.
This means that if we want to loop through the StreamReader
using the `read()` function, we must get the first block
via `StreamReader::currentBlock()` first.

2. The `block` variable needs to be updated for each iteration
in `createSTEMImage` so that `block.header.imagesInBlock`
returns the correct number.

The python example `stem_image.py` now appears to work.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>