Add option for zipping raw data
The raw data will not be zipped by default,
but there is at least an option to do so.

It also zips the data in chunks whose sizes are that
of the frame size. This allows for much quicker reading
of individual frames.

Also changed the path where the raw data will be saved
to be in `/stem/frames` instead of
`/electron_events/frames`.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>