Expand sparsified data in C++ instead of Python
This results in a very dramatic speed increase, from
~50 - 60 seconds on my computer to 0.5 seconds.

Timings seem to indicate that the vast majority of time
was spent in between the pybind11 function call and the start
of the C++ function. Whatever was happening in between (perhaps
a data copy or conversion) was taking a long time.

It is now much faster.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>