Use one mask and one image for create_stem_image
Before, the bright field mask would have an inner radius of 0
and an outer radius set by the user. Now, we will allow users to
just set all of the inner and outer radii themselves.

Because the user can now pass in a list of inner radii and outer radii,
it is reasonably efficient for users to be able to pass in all of the
radii of their own choosing.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>