Return dimensions from electronCount() function
A new struct, ElectronCountedData, was added that contains the previous
output of electronCount(), plus the scan and frame dimensions. This is
done so that the scan and frame dimensions can be obtained again later
in the code.

A new overload of createSTEMImagesSparse() was also added that takes in
an ElectronCountedData argument. This function can be called from the
python code as well.

An example of using the ElectronCountedData for creating a sparse stem
image can be seen now in `electron.ipynb`.

This fixes the third point of #76.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>