Exit with error if num processes exceed num files
Before this commit, if the number of processes exceeded the number of
files, some of the processes would have no files to read, and an
exception would be raised when it tried to read the list of empty files.

This adds an error check to make sure the number of processes do not
exceed the number of files.

Another alternative would be to allow the readers to run anyways (even
if there are no files) and to create empty datasets with them, and use
those in the reduction. But this might be tricky to do...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>