Do not expand sparse data for sparse stem images
When the sparse stem image functions were originally created, they
expanded the sparse data so that they could re-use functions that
are used by regular stem images. This, however, creates a very large
copy for some datasets.

This modification avoids that expansion and simplifies sparse stem
image creation. It should, in theory, be faster as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>