Use tuples and std::pair for dimensions and points
Rather than having separate variables for things such as width and
height, combine them into one variable. In python, this is a tuple, and
in C++, this is a std::pair.

This combines scan width and scan height into "scan_dimensions" (python)
and "scanDimensions" (C++) variables. Frame width and frame height are
combined into "frame_dimensions" (python) and "frameDimensions" (C++).

Center x and Center y are combined into just a "center" variable.

Typically, dimensions.first is the width, and dimensions.second is the
height. But there is one place where the ordering was switched, because
it appeared to be wrong: maximalPointsParallel in electron.cpp. This
needs to be tested with a non-square dataset.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>