Avoid copying data for PyReader's PyBlock
The data was previously copied in order to prevent seg faults. These
changes fix the seg faults, though, by keeping a copy of the pybind11
object around so that internal PyObject's reference count doesn't go
to zero.

A wrapper is needed to store the pybind11 object so that the API for
calling it is similar to std::shared_ptr, so that it can be used
in the stempy C++ template functions. Additionally, the pybind11 object
is stored in a std::shared_ptr internally, so that the PyBlock object
can be copied without copying the data (as is done with other types of
Block objects in stempy).

const_cast is needed in io.cpp because the buffer_info constructor
requires a non-const pointer. pybind11 also internally uses const_cast
to achieve the same purpose.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>