Don't check out-of-frame pixels in maximalPoints()
This adds some checks to make sure that, for border pixels, the pixels
are not compared to pixels outside the frame. For example, if a pixel
is on the top row of a frame, we cannot compare it to pixels above it.
It is unknown whether the prior behavior involved correct wrapping, or
something else.

Vtkm seems to wrap, but this probably doesn't matter since the border
pixels are most likely the least important ones.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>