Sort frame data when writing to an HDF5 file
In order for a reader of the HDF5 stempy file to know the
scan positions, one of two things needs to be true:

1. The frame images need to be sorted numerically, starting with 0
2. The "scan_positions" need to be written to a data set

This fixes the "create_hdf5.py" file so that the frame images are
sorted numerically (additionally, these changes also remove the need
to have two copies of the data in memory, which was a problem with
"create_hdf5.py" before).

This commit also adds an option to write the scan positions in
stempy.io.save_raw_data(), in case that is preferable to callers.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>