Allow np.ndarray input to be either raw or sparse
For create_stem_images(), allow numpy.ndarray input to be either
raw or sparse data. The type of data is inferred by the presence
of the "frame_dimensions" argument. If "frame_dimensions" was not
passed, it implies that the data is raw data. If it was passed, then
it implies that the data is sparse data.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>