Add ability to h5readwrite to write list attribute
Add the ability to write a list/vector to an attribute. This
will be done for dimensions attributes, as they will be written as
tuples.

We needed to make the std::string specialization call the c string
specialization instead of the other way around, because some compile
error would occur otherwise.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>