Ensure GIL is acquired before deleting py::array
Although the DataHolder class was acquiring the GIL before
decrementing the reference count on the shared pointer when "reset()"
was called, there were a few cases where the GIL was not acquired.
This causes a crash when the last shared pointer is destroyed.

The cases in DataHolder where the GIL was not acquired were:

1. Destructor
2. Copy assignment
3. Move assignment

Both 1 and 2 were being performed in the electron counting code.

This hopefully fixes all of the cases where the shared pointer's
reference count is decremented without the GIL being acquired.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>