Only read python attributes from electron counts
Rather than re-using the C++ object, generate the stem image
using only the python attributes from the ElectronCountedData.
This is done so that the C++ object is no longer required for
the electron counted data to be used in create_stem_images().

Now, if *any* object is passed to create_stem_images that
has 'data', 'frame_dimensions', and 'scan_dimensions' attributes,
it can be used as the electron counted data.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>