Add sparse slicing to SparseArray
If sparse slicing is set (default is True), then when slicing is
performed, and the result will not be a single frame, then a new sparse
array will be returned from the slicing operation instead of a dense array.

This allows for things like cropping/striding the data while keeping the
array sparse.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>