Count duplicates in sparse array
Previously, we were assuming that the presence of a sparse value
in a frame indicated that the value would be 1.

However, if there were 2 instances of the sparse value in the frame,
then the value should be 2. As in, we should have been counting
the number of occurrences of the sparse values in the frame.

Although this hurts performance, we are now counting the number of
occurrences of sparse values in each frame, so that we can give an
accurate answer. This will also be important for when we add in
binning, since it also uses duplicates in the sparse array.

Some of the results including the max() are now different.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>