Require cmake 3.19
CMake 3.18 is required for the use of "Development.Module" in
FindPython. This is needed to avoid finding the python libraries when
they are not needed, such as for building wheels.

CMake 3.19 is required for the "hdf5::hdf5" cmake target.

We can backport both of these if we decide we want to support an
earlier version of cmake.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>