Export all symbols on windows
Windows does not export and import symbols by default, which makes it
so our python modules can't import them.

A classic way to solve this is by putting __declspec(dllimport)
and __declspec(dllexport) into the source code, which can be done via
a macro generated by CMake and its GenerateExportHeader module. The
other classic way is to create .def files with all the symbols in it
that you want exported.

The newer, easier way to solve this is by simply passing
`-DCMAKE_WINDOWS_EXPORT_ALL_SYMBOLS=TRUE` to cmake. CMake will then
automatically create .def files. For some reason, setting this option
in the CMakeLists.txt file does not work. I have to pass it via the
command line. So I am adding it to the options in the setup.py file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>