Add testing tempfile workarounds for Windows
Unfortunately on Windows, you can't open the same file twice in
a row. So we have to avoid the tempfile context manager, which opens
the file, and we have to keep the file from being deleted automatically
when it gets closed.

Then we have to delete it ourselves when we are done. It's a pain...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>