Add requirements.txt, and use this in setup.py
It would be nice if either:

1. pip supported installing dependencies only from setup.py
2. There was an offical way to load requirements.txt from setup.py

So we don't have to duplicate requirements.

However, just opening and reading the setup.py file seems to mostly
work. So we can try to do that for now...

We are doing this because we need a way to install only dependencies,
and not the whole package.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>