Skip pip install stempy for read the docs
It isn't really needed - we are already appending the path
to the python files to the sys.path. And we are mocking the
stempy._io and stempy._image imports already.

This also adds _threaded_multi_pass_reader to the list of _io_mock
objects to prevent an error involving multiple inheritance with
different metaclasses.

deprecation is also added as a requirement for the docs, since it
is needed to display the deprecation in the documentation.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>