Dereference symlinks to fix singularity
Previously, symlinks were being preserved when the copy occurred. The
symlinks pointed into the /home/jovyan directory.

When a singularity shell was started, the user would be "jovyan", so
there would be no issue importing the files.

However, if jupyter was started instead, the user would be the current
user on the host OS. And this user would have permissions issues with
reading into the /home/jovyan directory.

To fix this, dereference the symlinks on `cp` to avoid having symlinks
pointing into `/home/jovyan`.

This was tested and seems to have fixed the issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>