Separate sparse arrays with same scan position
Previously, in the electron counting, we would combine two sparse arrays
if they had the same scan positions. Now, we are modifying this so that
the sparse arrays are kept separately per frame.

The ElectronCountedData struct now has a scanPositions vector to indicate
which scan a sparse array belongs to, since the data vector no longer
matches the size of the scan.

This also updates the SparseArray class to perform faster (and correct)
arithmetic and expanding based upon the new way that the sparse data is
formatted.

I have tested the electron counting with a dataset that has two frames
per scan position, and for fully expanding the output sparse array, I
received the same result before and after these changes, which indicates
that it is hopefully working correctly.

An extensive test was also added which tests most pieces of using a
SparseArray where some scan positions have multiple frames.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>