Add metadata to SparseArray
This adds metadata to the SparseArray class than will be saved/loaded
via HDF5. A basic test was also added to ensure this works.

Now, when electron counting is performed, the metadata on the sparse
array gets set to the electron counting metadata (thresholds mostly).
This way, we can also keep and save the electron counting metadata to
the HDF5 file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>