Refactor SparseArray and methods to data version 3
Version 3 of the SparseArray and electron counting methods changes the
data shape from a flat array of sparse frames to a 2D array, where the
first dimension is the scan position, and the second dimension is the
index of a frame at that scan position. This removes the need for the
`scan_positions` array (previously used to indicate which scan position
each sparse frame corresponded to).

Now, the data array passed to the SparseArray must either have a shape
of `(num_scans, frames_per_scan)` or just `(num_scans,)` (where one frame
per scan will be assumed, and the shape will be changed automatically to
`(num_scans, 1)`. Validation is performed to ensure this is the case, or
an exception will be raised.

This simplifies a lot of the logic involving multiple frames per scan
position, and completely removes the need for manipulating the scan
positions array that the SparseArray previously held.

For writing to HDF5, it appears that h5py does not allow a 2D array of
variable length arrays to be written (see h5py/h5py#876), so we must
flatten it to a 1D array for HDF5 storage. The scan shape allows the
2D array to be re-created on loading.

Backwards compatibility has been maintained so that the SparseArray class
can load in HDF5 files saved in versions 1, 2, or 3 and correctly convert
them to version 3 upon loading. Tests have been added to ensure this is
the case.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>