Fix indexing issue for electron counting
This was using the old electron counting method that we don't use anymore
except when we are using an HDF5 file, and only *without* VTKm. However,
it is a bug that needs to be fixed. The `i` index is for the frame number,
but the `j` index is for the index inside the frame. These should be the `j`
index instead.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>