Simplify slicing more by splitting scan and frame
We are now really forwarding everything involving indexing to numpy
that we can, and allowing numpy to take care of what the shape should
be, and where the resulting indices should end up.

This also now splits up the scan and frame slices more formally, and
adds a lot of validation with human-readable errors.

More tests were also added which validate nearly everything that we allow.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>